backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mongoengine>=0.27
anthropic>=0.25
stripe>=8.0
reportlab>=4.0
python-pptx>=0.6
cloudinary>=1.36
django-cors-headers>=4.3
python-dotenv>=1.0
dnspython>=2.4
pymongo>=4.6
requests>=2.31
Pillow>=10.0
